Removed remaining redundant datapack information
It is bad practice to include all data files from another mod, it makes the repo harder to maintain, more likely to break with mod updates, and risks legal issues depending on the original mod's license. No content changes were made in this commit as all of this data is already present in the source mods.
Mods affected
- Beyond Earth (mod was previously removed)
- Cooking for Blockheads
- Ender Storage
- Ex Deorum
- Mekanism Generators
- Pam's Harvestcraft 2: Crops
- Pneumaticcraft
- Special Mobs
- Survive
- Sync
- Torchmaster