Survive Update (see description)
THIS UPDATE IS BROKEN:
The newest version of UnionLib is broken and this change will be immediately reverted after pushing. When UnionLib is fixed, I will push the commit again, but I want to push this now to preserve the changes and log.

Updated Survive
- Fixes thirst damage hurting armor
- Obstacles blocking heat sources now affect temperature
Updated UnionLib
Forge should now be 47.4.0